refactor: javascript purge
