+++ e2e/plugins/index.ts
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-const cucumber = require('cypress-cucumber-preprocessor').default;
+import cucumber from 'cypress-cucumber-preprocessor';
 
 module.exports = (on: Function) => {
   const options = {
