+++ package.json
@@ -52,6 +52,7 @@
     "@storybook/addon-links": "^6.0.22",
     "@storybook/addon-storysource": "^6.0.22",
     "@storybook/react": "^6.0.22",
+    "@testing-library/react": "^11.0.4",
     "@types/cypress-cucumber-preprocessor": "^1.14.1",
     "@types/jest": "^26.0.14",
     "@types/node": "^14.11.2",
